// api/generated/model/genderEnum.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Backend API
 * API built from the django-backend-template.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `male` - Male
 * * `female` - Female
 */
export type GenderEnum = typeof GenderEnum[keyof typeof GenderEnum];


export const GenderEnum = {
  male: 'male',
  female: 'female',
} as const;